Debounce updating take-home pay results due to input change
Trying to update this instantly was causing too many concurrent updates, especially on mobile when sliding the income slider on Chrome. That would result in a React error and a blank page.
This debounces with a 50 ms wait to update results. It's noticeably different but not slow feeling. Especially on mobile, the user probably can't see the results on the same screen while sliding anyway.