Update residence tax tooltip
There was no mention before about the fact residence tax is billed in arrears, meaning it is based on your previous year's income and the payments don't necessarily align with the calendar year. Give a more thorough description in the tooltip in the hopes this draws enough attention to this peculiarity of residence tax. It may be worth taking other measures to draw attention to this.