Implement show calculation details for taxes (#16)
Enables the "Show Calculation Details" toggle and includes some details in calculating taxes for now. This can be expanded in the future to include details about the social insurance calculations. This also fixes the calculation of the net employment income for gross employment incomes under 6.6 million, which was missing some nuances of that calculation (e.g., rounding down to the nearest 4000 yen in some cases).