Account for adjustment deduction in residence tax
The residence tax calculation did not apply the adjustment deduction (調整控除) at all before. It is a small difference, but for accuracy the calculation now includes this. The only applicable deduction that is factored into this that the calculator supports is the basic deduction. In the future, when more deductions are supported, the logic for calculating the adjustment deduction will need to be updated.