Include percentage of each result
Space-permitting, include the percentage of each result row since people sometimes perfer to refer to percentages rather than absolute values. On mobile, the percentages will not be shown in the results component because it causes word wrapping due to limited horizontal space. The chart tooltip will have the percentages, however, regardless of mobile or not.